requirements.txt
cirq-core
matplotlib
sympy
scipy
pyqsp